Oh my. Too many changes in one commit. Incremental tests/commits could have been started earlier. Still have a bit more to do, but it's a bit trickier, at least from an early impression, than the others, which seemed less tricky, but weren't always.
To the package proper, a text element was added, and the old Quadrant
api was separated from the NotTurtle api. Both remain in one module,
renamed to "lib". Updated the "old" Quadrant sample to conform.

Renamed the sample files to be less obtusely generic.

Triphi sample:
  Much more organized and reusable. Also bigger! Where two files were,
  now there are seven, although two are short stubs. Pulled some out
  into a reusable package, too!
  Added a layer with labels for the vertices of the triangles. Uses that
  new Text element!
  Rotated the labels so they line up more intuitively.
  Flip option for a mirrored rendering.
  More math:
    Given an iterable of points and a center, sort the points as though
    they are points on the circle, by increasing angle, or counter/anti-
    clockwise.
    Given a line segment as one side, find the two points that define
    the equilateral triangles that share that side. May also be used to
    find a pair of isosceles triangles. Sorts the two triples counter-
    clockwisely, to conform with convention.

Made a minor commentary change to the text file with the math worked
out. Noticed that there's probably more clarification neeeded, some
labels are wrong, and maybe there's the same results worked out three
ways, of which two should get out of there and into a supplementary
file, if retained at all.