"figure" sample had some glitches/incompletenesses; geometry updates on the command line were not applied, and all layers were invisible if any options were given. That was left over from the old, super-basic command line parsing.
Tiles are done! At least the bit that makes a tiling. Not the
interesting bit that extracts a rectangular unit from a tiling.

The flipped figure wasn't. A lot of sad code is happier now.

rotated point labels again. quit that!

Containers are interesting. Not sure I like the attrdictbase type after
all. The full attrdict, if it was ready, might be alright. Restored the
old container, added support for the LazyEval type for immutable mutable
defaults, and an update() method.

The replacement wasn't quite integrated, either. The stroke-width
scaling wasn't working. Oops!