Broke up the main module into a subpackage a a bit mess of little modules. Might consolidate those a bit.
Back to no dependecies. The standard library is used under the lib
subpackage, but core/ has no absolute imports! That funny business that
wasn't backported to 2.7 is gone, and it's backward compatible with the
demo/sample and not too unpleasant.

There is an unused placeholder under core/ where a standard library
feature could fit, but I'll roll a replacement that is more explicitly
suited.