Thinking about importing XML (not parsing, ick!), which led to the same issue of copying: naked instantiation, with finished attributes. Figured out how pickle does it: __new__().
Copying got a lot of updating, mostly wrt hooks for customization, which
was necessary for a few classes, not just theoretical application
elsewhere/when/by otheres. The Path element hadn't gotten revised for
the initial stab at copy, but it should be good now.

A result of which is that Path can initialize off its attributes, and
the initial position is optional when there's a valid path there.

Shortened a couple method names. More consistent, and the calls will
shortly become longer.

Moved the export-to-ElementTree one-liner to a separate module, where
import code can go.