Ditched StyledElement; needed somthing more like a mixin, but not. Using flags, which can be extended by the user.
I found out the xml declaration hasn't got arbitrary attributes, so I
didn't need my twice-invoked utility there. I did reuse (once) the
utility that only had a single use. Balance!

The __str__() methods that produce the output are less messy. Whitespace
in the output got tweaked, that's going to change tests.

Renamed Clip class, and the corresponding method in stack class to
ClipPath, to be consistent with SVG. Every other classname is. Well,
except Group/g, but that one really needs it.

A couple Elements that took styles but weren't flagged/subclassed for it
are now.
Support for stylesheets.

A few more bits of import support.