Kinda bloopered with a quick shifting around a little code last commit, didn't test. Left out some quotes that kinda needed to be there!
Reworked the tag registry. A couple of containers and a function got
rolled up into a class instance, with some helpful methods added. Nice
to have fewey symbols cluttering up the namespace!

There was a PY2 symbol that wasn't really needed. The tests are at class
constructioni and might as well be direct checks of the feature in
question. Simplified the code under the checks, too.

The class attributes set at registeration are better enforced as not
being instance attributes.