style sheet support was read-only! filled the rest in.
added a CSS class that is a Style element with kwargs turned into css
directives. key is the selector, value is a dictionary of key-value
pairs. The current implementation adds a lot of newlines that probably
will become optional.

Also a bunch of whitespace and commentary updates that lower the signal
to noise ratio!