Removed element registry bits.
Layer class was delegating up three levels rather than to parent, which was
functionally identical and presumably saved some time in method resolution,
but would not play well with overloading in a subclass. Uses parent now, which
will be easier to shift into a "sane employment" of super() eventually.

"Element" and "Style" inherited from built-in "dict". Now they inherit from a
type that is uniform across Py2 and Py3, emulating the newer form if executed
under Py2.

Tweaked the Style/CSS element class to have a better API.